https://yt.iteco.dev/issue/ISPP-1153

+++ processor/src/web/java/ru/axetta/ecafe/processor/web/internal/FrontController.java
@@ -2807,7 +2807,7 @@ public class FrontController extends HttpServlet {
                 return new DocumentCreateResponse(DocumentCreateResponse.ERROR_CLIENT_IS_STUDENT,
                         DocumentCreateResponse.ERROR_CLIENT_IS_STUDENT_MESSAGE);
             }
-            boolean saveToMk =  ClientManager.isClientGuardian(persistenceSession, client);
+            boolean saveToMk = ClientManager.isClientGuardian(persistenceSession, client);
             DulDetail dulDetail = fillingDulDetail(persistenceSession, documentItem);
             dulDetailService.validateDul(persistenceSession, dulDetail, true);
             MeshDocumentResponse meshDocumentResponse = dulDetailService
@@ -2868,7 +2868,7 @@ public class FrontController extends HttpServlet {
                 return new DocumentUpdateResponse(DocumentUpdateResponse.ERROR_CLIENT_IS_STUDENT,
                         DocumentUpdateResponse.ERROR_CLIENT_IS_STUDENT_MESSAGE);
             }
-            boolean saveToMk =  ClientManager.isClientGuardian(persistenceSession, client);
+            boolean saveToMk = ClientManager.isClientGuardian(persistenceSession, client);
             dulDetailService.validateDul(persistenceSession, dulDetail, true);
             MeshDocumentResponse meshDocumentResponse = dulDetailService
                     .updateDulDetail(persistenceSession, dulDetail, client, saveToMk);
@@ -2922,7 +2922,7 @@ public class FrontController extends HttpServlet {
                 return new DocumentDeleteResponse(DocumentDeleteResponse.ERROR_CLIENT_IS_STUDENT,
                         DocumentDeleteResponse.ERROR_CLIENT_IS_STUDENT_MESSAGE);
             }
-            boolean saveToMk =  ClientManager.isClientGuardian(persistenceSession, client);
+            boolean saveToMk = ClientManager.isClientGuardian(persistenceSession, client);
             MeshDocumentResponse meshDocumentResponse = dulDetailService
                     .deleteDulDetail(persistenceSession, dulDetail, client, saveToMk);
             if (!meshDocumentResponse.getCode().equals(MeshDocumentResponse.OK_CODE)) {
@@ -3097,7 +3097,7 @@ public class FrontController extends HttpServlet {
 
             Client child = DAOReadonlyService.getInstance().findClientsByMeshGuid(childMeshGuid);
             if (child == null) {
-                return new GuardianMeshGuidResponse(ResponseItem.ERROR_INTERNAL, ResponseItem.ERROR_INTERNAL_MESSAGE);
+                return new GuardianMeshGuidResponse(ResponseItem.ERROR_CLIENT_NOT_FOUND, ResponseItem.ERROR_CLIENT_NOT_FOUND_MESSAGE);
             }
 
             personResponse = getMeshGuardiansService().createPersonWithEducation(persistenceSession, idOfOrg, firstName, patronymic, lastName, genderId, birthDate, snils,
@@ -3233,8 +3233,6 @@ public class FrontController extends HttpServlet {
             @WebParam(name = "idOfOrg") Long idOfOrg,
             @WebParam(name = "informing") Boolean informing) {
 
-        //todo для чего idOfOrg?
-
         Session persistenceSession = null;
         Transaction persistenceTransaction = null;
         try {
@@ -3255,22 +3253,37 @@ public class FrontController extends HttpServlet {
             criteria.add(Restrictions.eq("meshGUID", childMeshGuid));
             Client child = (Client) criteria.uniqueResult();
 
-            MessageContext mc = wsContext.getMessageContext();
-            HttpServletRequest req = (HttpServletRequest) mc.get(MessageContext.SERVLET_REQUEST);
-            ClientGuardianHistory clientGuardianHistory = new ClientGuardianHistory();
-            clientGuardianHistory.setReason("Создана связка арм методом \"addGuardianToClient\"");
-            clientGuardianHistory.setWebAdress(req.getRemoteAddr());
+            if (guardian == null || child == null) {
+                return new GuardianResponse(GuardianResponse.ERROR_CLIENT_NOT_FOUND,
+                        GuardianResponse.ERROR_CLIENT_NOT_FOUND_MESSAGE);
+            }
+            ClientGuardian clientGuardian = getClientGuardian(persistenceSession, guardian, child);
 
-            Long newGuardiansVersions = ClientManager.generateNewClientGuardianVersion(persistenceSession);
-            ClientManager.addGuardianByClient(persistenceSession, child.getIdOfClient(), guardian.getIdOfClient(), newGuardiansVersions,
-                    true, ClientGuardianRelationType.fromInteger(relation), ClientManager.getNotificationSettings(),
-                    ClientCreatedFromType.ARM, ClientGuardianRepresentType.fromInteger(typeOfLegalRepresent), clientGuardianHistory,
-                    ClientGuardianRoleType.fromInteger(agentTypeId), !informing);
+            if (clientGuardian == null) {
+                MessageContext mc = wsContext.getMessageContext();
+                HttpServletRequest req = (HttpServletRequest) mc.get(MessageContext.SERVLET_REQUEST);
+                ClientGuardianHistory clientGuardianHistory = new ClientGuardianHistory();
+                clientGuardianHistory.setReason("Создана связка арм методом \"addGuardianToClient\"");
+                clientGuardianHistory.setWebAdress(req.getRemoteAddr());
 
-            MeshAgentResponse personResponse = getMeshGuardiansService().addGuardianToClient(meshGuid, childMeshGuid, agentTypeId);
-            if (!personResponse.getCode().equals(PersonResponse.OK_CODE)) {
-                logger.error(String.format("Error in addGuardianToClient %s: %s", personResponse.getCode(), personResponse.getMessage()));
-                return new GuardianResponse(personResponse.getCode(), personResponse.getMessage());
+                Long newGuardiansVersions = ClientManager.generateNewClientGuardianVersion(persistenceSession);
+                ClientManager.addGuardianByClient(persistenceSession, child.getIdOfClient(), guardian.getIdOfClient(), newGuardiansVersions,
+                        true, ClientGuardianRelationType.fromInteger(relation), ClientManager.getNotificationSettings(),
+                        ClientCreatedFromType.ARM, ClientGuardianRepresentType.fromInteger(typeOfLegalRepresent), clientGuardianHistory,
+                        ClientGuardianRoleType.fromInteger(agentTypeId), !informing);
+
+                MeshAgentResponse personResponse = getMeshGuardiansService().addGuardianToClient(meshGuid, childMeshGuid, agentTypeId);
+                if (!personResponse.getCode().equals(PersonResponse.OK_CODE)) {
+                    logger.error(String.format("Error in addGuardianToClient %s: %s", personResponse.getCode(), personResponse.getMessage()));
+                    return new GuardianResponse(personResponse.getCode(), personResponse.getMessage());
+                }
+            }
+            if (!idOfOrg.equals(guardian.getOrg().getIdOfOrg()) && MigrantsUtils.findActiveMigrant(
+                    persistenceSession, idOfOrg, guardian.getIdOfClient()) == null) {
+                Org org = persistenceSession.get(Org.class, idOfOrg);
+                Date date = new Date();
+                ClientManager.createMigrationForGuardianWithConfirm(persistenceSession, guardian, date, org,
+                        MigrantInitiatorEnum.INITIATOR_ORG, VisitReqResolutionHistInitiatorEnum.INITIATOR_CLIENT, 12);
             }
 
             persistenceTransaction.commit();
@@ -3310,21 +3323,26 @@ public class FrontController extends HttpServlet {
             criteria.add(Restrictions.eq("meshGUID", childMeshGuid));
             Client child = (Client) criteria.uniqueResult();
 
-            MessageContext mc = wsContext.getMessageContext();
-            HttpServletRequest req = (HttpServletRequest) mc.get(MessageContext.SERVLET_REQUEST);
-            ClientGuardianHistory clientGuardianHistory = new ClientGuardianHistory();
-            clientGuardianHistory.setReason("Удалена связка арм методом \"deleteGuardianToClient\"");
-            clientGuardianHistory.setWebAdress(req.getRemoteAddr());
+            if (guardian == null || child == null) {
+                return new GuardianResponse(GuardianResponse.ERROR_CLIENT_NOT_FOUND,
+                        GuardianResponse.ERROR_CLIENT_NOT_FOUND_MESSAGE);
+            }
+            ClientGuardian clientGuardian = getClientGuardian(persistenceSession, guardian, child);
 
-            Long newGuardiansVersions = ClientManager.generateNewClientGuardianVersion(persistenceSession);
+            if (clientGuardian != null) {
+                MessageContext mc = wsContext.getMessageContext();
+                HttpServletRequest req = (HttpServletRequest) mc.get(MessageContext.SERVLET_REQUEST);
+                ClientGuardianHistory clientGuardianHistory = new ClientGuardianHistory();
+                clientGuardianHistory.setReason("Удалена связка арм методом \"deleteGuardianToClient\"");
+                clientGuardianHistory.setWebAdress(req.getRemoteAddr());
 
-            ClientManager.removeGuardianByClient(persistenceSession, child.getIdOfClient(), guardian.getIdOfClient(),
-                    newGuardiansVersions, clientGuardianHistory);
-
-            MeshAgentResponse personResponse = getMeshGuardiansService().deleteGuardianToClient(agentMeshGuid, childMeshGuid);
-
-            if (!personResponse.getCode().equals(PersonResponse.OK_CODE)) {
-                return new GuardianResponse(personResponse.getCode(), personResponse.getMessage());
+                Long newGuardiansVersions = ClientManager.generateNewClientGuardianVersion(persistenceSession);
+                ClientManager.removeGuardianByClient(persistenceSession, child.getIdOfClient(), guardian.getIdOfClient(),
+                        newGuardiansVersions, clientGuardianHistory);
+                MeshAgentResponse personResponse = getMeshGuardiansService().deleteGuardianToClient(agentMeshGuid, childMeshGuid);
+                if (!personResponse.getCode().equals(PersonResponse.OK_CODE)) {
+                    return new GuardianResponse(personResponse.getCode(), personResponse.getMessage());
+                }
             }
             persistenceTransaction.commit();
             persistenceTransaction = null;
@@ -3410,6 +3428,15 @@ public class FrontController extends HttpServlet {
         dulDetail.setExpiration(documentItem.getExpiration());
         dulDetail.setLastUpdate(currentDate);
         return dulDetail;
+    }
+
+    private static ClientGuardian getClientGuardian(Session persistenceSession, Client guardian, Client child) {
+        Criteria criteria;
+        criteria = persistenceSession.createCriteria(ClientGuardian.class);
+        criteria.add(Restrictions.eq("idOfChildren", child.getIdOfClient()));
+        criteria.add(Restrictions.eq("idOfGuardian", guardian.getIdOfClient()));
+        criteria.add(Restrictions.ne("deletedState", true));
+        return (ClientGuardian) criteria.uniqueResult();
     }
 
 }
